+++ tic-tac-toe/src/app/components/textInput/index.js
@@ -1,8 +1,8 @@
 import React, { Fragment } from 'react';
 
-export const CustomInput = props => (
+export const textInput = props => (
   <Fragment>
-    <label>{props.label}</label>
+    <label htmlFor="label">{props.label}</label>
     <input {...props.input} type={props.type} />
   </Fragment>
 );
