+++ tic-tac-toe/src/app/screens/Game/components/Board/index.js
@@ -10,7 +10,6 @@ class Board extends Component {
   }
 
   render() {
-    
     return (
       <Fragment>
         <div className={styles.boardRow}>
