+++ tic-tac-toe/src/app/screens/Login/fields.js
@@ -1,0 +1,5 @@
+export const FORM_NAME = 'login';
+export const FIELDS = {
+  user: 'user',
+  pass: 'pass'
+};
