+++ tic-tac-toe/src/app/screens/Login/index.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import Login from './layout';
 
 class LoginContainer extends Component {
-  submit = values => alert(JSON.stringify(values, null, 4));
+  submit = values => alert(JSON.stringify(values));
 
   render() {
     return <Login onSubmit={this.submit} />;
