+++ tic-tac-toe/src/index.js
@@ -1,16 +1,15 @@
 import React from 'react';
+import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
 
 import './scss/index.scss';
-import { Provider } from 'react-redux';
-
 import store from './redux/store';
 import registerServiceWorker from './registerServiceWorker';
-import Login from './app/screens/LoginContainer';
+import LoginContainer from './app/screens/Login';
 
 ReactDOM.render(
   <Provider store={store}>
-    <Login />
+    <LoginContainer />
   </Provider>,
   document.getElementById('root')
 );
