--- tic-tac-toe/src/redux/gameReducer/reducers.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import { reducer as formReducer } from 'redux-form';
-
-const reducers = {
-  form: formReducer
-};
-
-const reducer = combineReducers(reducers);
-
-export default reducer;
