+++ tic-tac-toe/src/redux/store.js
@@ -1,6 +1,11 @@
-import { createStore } from 'redux';
+import { createStore, combineReducers } from 'redux';
+import { reducer as formReducer } from 'redux-form';
 
-import reducer from './gameReducer/reducers';
+const reducers = {
+  form: formReducer
+};
+
+const reducer = combineReducers(reducers);
 
 const store = createStore(
   reducer,
